+++ agent/dist-desktop/core/collector/douyin/network/interceptor.js
@@ -1,5 +1,0 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-exports.DouyinNetworkInterceptor = void 0;
-var response_collector_js_1 = require("../../../network/response-collector.js");
-Object.defineProperty(exports, "DouyinNetworkInterceptor", { enumerable: true, get: function () { return response_collector_js_1.ResponseCollector; } });